Wire up auditbeat, heartbeat, osquerybeat and packetbeat OTel receivers
The four beat receiver types added in PR #14029 were registered as
collector components but never wired into the translation layer, so
GetOtelConfig would return an error for any component backed by one
of these beats.

Add the missing cases in three switch statements in otelconfig.go:

- getSignalForComponent: auditbeat, heartbeat, osquerybeat, packetbeat
  all use pipeline.SignalLogs (same as filebeat).
- getReceiverTypeForComponent: map each beat binary name to its
  receiver type: auditbeat→abreceiver, heartbeat→hbreceiver,
  osquerybeat→osqreceiver, packetbeat→pbreceiver.
- getDefaultDatastreamTypeForComponent: all four produce "logs"
  datastreams (same as filebeat).

Fix the receiver-config switch in getReceiversConfigForComponent to
key on the beat binary name (binaryName, from comp.BeatName()) rather
than the truncated receiver-type prefix (which was only correct for
filebeatreceiver/metricbeatreceiver). Add the beat-specific config
key for each new receiver:
  auditbeat.modules, heartbeat.monitors, osquerybeat.inputs, packetbeat.protocols

Also correct the test's expected packetbeat port values from int to
float64, matching how numbers round-trip through the confmap pipeline.

+++ internal/pkg/otel/translate/otelconfig_test.go
@@ -602,7 +602,7 @@ func TestGetOtelConfig(t *testing.T) {
 					"data_stream": map[string]any{
 						"dataset": "generic-1",
 					},
-					"ports":      []any{443, 8443},
+					"ports":      []any{float64(443), float64(8443)},
 					"index":      "logs-generic-1-default",
 					"processors": defaultInputProcessors("test-1", "generic-1", "logs"),
 					"type":       "packet",
